kvm: handle fs and gs with cpl > 0
vmx imposes the quite unnecessary restriction that the host fs and gs selectors
have cpl = 0 and ti = 0.  however, it is quite legal to have cpl != 0 or
ti = 1 for fs and gs.  this means that we must artificially set the host fs
and gs fields to zero, and restore the segments manually.

an additional complication is that loading gs in long mode will overwrite
the gs base register which is the PDA.  so we need to restore that too.  yuck.