kvm: avoid longjmp() as a method of existing kvm_run()
kvm_run() loops intenally to service exits.  this allows a simple callback
model and reduces register conversions, but sometimes we need to exit the loop.
we did this by calling cpu_loop_exit(), which is just a longjmp().

unfortunately, this is problematic on 32-bit because longjmp() clobbers
registers, and qemu tries to store a lot of global variables in registers.

so, avoid that by having the callbacks return a value, and terminate the loop
if the value is nonzero, also returning that value.  this avoids the longjmp()
and is easier to understand.