kvm: fix host memory corruption
in the page fault path, we mark a page header's dirty slot bitmap so we can
track dirty pages.  however, we call the update function even for I/O pages,
which don't have a memory slot.

this causes a random bit in memory to be set.  on x86_64, this was lost in
the unused memory on the machine.  luckily, on i386 it oopsed.

fix by calling the bitmap update function only after we're sure the pte maps
non-I/O memory.