kvm single-qemu-binary: replace all env->use_kvm with kvm_allowed in cirrus_vga.c
Also removed some appropriate "#ifdef USE_KVM" and "env" declarations.
cirrus_vga should not tied to a specific cpu, but to the global kvm_allowed.
Furthermore, there are incidents where env is NULL.

The title "kvm single-qemu-binary" above also fits the previous
        commits: r3480, r3612.
Its purpose is to enable single qemu executable that is able to run
        with/without kvm kernel module support.
Single-qemu-binary was introduced by Anthony Liguori on the kvm-devel list.
Mr Liguori was notified about the problem fixed in this commit and
        would probably publish a fix of his own in the near future.