qemu savevm: save/load usb-uhci (#154)
This includes "usbdevice tablet".

from Lonnie Mendez on the qemu-devel mailing list (adjusted to qemu 0.8.2):
http://lists.gnu.org/archive/html/qemu-devel/2006-11/msg00239.html
It is mentioned that "The code will only work well if you remove all usb
devices attached to the guest prior to saving the vm state."
It was working for me even without removing, but need more testing.