From Markus Rechberger <markus.rechberger@amd.com>:
This patch is a sync with the latest apic code in the qemu repository, this fixes a system crash after linux tries to set up apic after the first reboot.
Signed-off-by: Markus Rechberger <markus.rechberger@amd.com>

I only changed the apic version_id the register_vmsave so the apic_timer will be serialized.
After the change the apic timer works after vm save/load cycle.
Was also sent to Qemu devel list.