kvm: vmx: Fix register constraint in launch code
From: Herbert Xu <herbert@gondor.apana.org.au>

Both "=r" and "=g" breaks my build on i386:

   $ make
   CC [M]  drivers/kvm/vmx.o
   {standard input}: Assembler messages:
   {standard input}:3318: Error: bad register name `%sil'
   make[1]: *** [drivers/kvm/vmx.o] Error 1
   make: *** [_module_drivers/kvm] Error 2

The reason is that setbe requires an 8-bit register but "=r" does not
constrain the target register to be one that has an 8-bit version on
i386.

According to

   http://gcc.gnu.org/bugzilla/show_bug.cgi?id=10153

the correct constraint is "=q".

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

[pulled from git]