Report the host's cpu vendor to the guest
x86_64 Linux uses the vendor ID as a means to select between sysenter and
syscall as the vdso kernel entry instruction for comaptibility mode (32-bit
userspace).  Because syscall is not supported in Intel's implementation of
compatibility mode, and sysenter is not supported in AMD's implementation of
compatibility mode, we can't have a least-common-denominator setting.

This is bad news for inter-vendor migration, at least for 32-on-64 setups.
Maybe we can trap the #UD and emulate the offending instruction.  Yuck.