Make sure to sync registers in monitor/gdbstub (v2)
This patch adds support to gdbstub and the monitor to sync CPU state before
and after the user views/changes it.  I also added kvm_allowed guards to
kvm_{load,save}_registers.  I chose to do it here so that callers didn't need
to use the guards.  We should probably provide nop implementations so that
#ifdef USE_KVM guards aren't required either.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>