Work around silly timing loop in qemu bios
when the qemu bios detects the number of cpus, it uses a timing loop to
give all cpus a chance to enumerate themselves.  However, the loop count
is tuned to the qemu cpu characteristics, which are different from real or
virtual cpus.

so sleep a bit after the first sipi to let the other cpus pick up.