In-kernel LAPIC live migration support
This patch adds the live migration support for in-kernel LAPICs. The
kernel and qemu have different data structures, kernel saves the apic to the
userspace as a apic page while qemu uses separate varaibles. The translation
is done in this patch. Like PIC/IOAPIC, the save/restore uses original qemu
savevm.

Signed-off-by: Yaozu (Eddie) Dong <eddie.dong@intel.com>
Signed-off-by: Qing He <qing.he@intel.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>