qemu: fix a possible deadlock situation in save/restore
kvm: qemu: fix a possible deadlock situation in save/restore

If a VM using userspace irqchip is saved when it's in HLT state,
and is restored to use the in-kernel irqchip, the HLT state flag
is also restored. This HLT flag in userspace HLT handling prevents
kvm kernel driver from being run. And at the same time, the original
userspace irqchip is now unused thus will not generate any interrupt
to wake up the HLT state. This patch fixes this deadlock situation
by dropping the HLT state at restore time

Signed-off-by: Qing He <qing.he@intel.com>
Signed-off-by: Yaozu (Eddie) Dong <eddie.dong@intel.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>