Make kvm sleep if eflag is unset when hlt is executed
In case the cpu is in a halt state, and there are pending interrupts, kvm
usually enters guest mode immediately to service interrupts.  But if interrupts
are disabled, this is pointless and we end up spinning.

Fix by sleeping in case we're in halt state with interrupts disabled.

This fixes spinning on the boot loader when the disk is damanged.

Signed-off-by: Dor Laor <dor.laor@qumranet.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>