kvm: libkvm: make kvm_run() more arch friendly
This patch creates hook kvm_arch_run in kvm_run function. This
is to handle cases that are architecture specific on a run exit.
Currently there is only one case that is handled that is obvious.
Others have at least some similarity on all archs.

Though really these return codes will need to be worked out in the
kernel kvm module. But here is a start.

Signed-off-by: Jerone Young <jyoung5@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>