kvm: libkvm: get_slot(): fix is-address-in-range check
For any slot the slot's allowed addresses are from start to (start+size-1).
The address start+size is outside of the slot's address range.

This fixes migration for 32 bit hosts. One of the
slots (the bios) is mapped at the top of the address space and its start+size
is 0. For that slot get_slot() failed.

Signed-off-by: Uri Lublin <uril@qumranet.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>