kvm: external module: add true sched_in preempt notifier emulation
Instead of hacking the module source to check whether the preempt notifier
sched_in hook needs to be called, add a breakpoint on the caller of
schedule() which triggered sched_out.  Use the Linux support for hardware
breakpoints to load the breakpoint when the process is switched to.

This greatly reduces the amount of hackery needed to support preempt
notifiers.

Signed-off-by: Andrea Arcangeli <andrea@suse.de>