Report revision 3 of the ACPI/SMBus PIIX4 controller
The PIIX4 ACPI controller prior to revision 0x3 contains a bug where
reading of the timer port is unreliable, so the kernel reads it three
times for consistency check.

QEMU does not suffer from that problem :)

The datasheet for PIIX4, PIIX4E, and PIIX4M is the same. I failed to
find any indication that the revision increase could affect anything
other than the PMTimer port read.

This reduces idle guest CPU consumption from 14% to 8% on 4-way KVM
guest.

Signed-off-by: Avi Kivity <avi@qumranet.com>