Remove hack that forced pci interrupts to be enabled
Previously, the acpi sci interrupt was implemented as a pci interrupt.
However, guests did not now about it and disabled the interrupt in the
routing registers, so a hack was introduced to force-enable the irqs.

Now that sci is implemented as an isa irq, this is no longer necessary,
so remove the hack.

Signed-off-by: Avi Kivity <avi@qumranet.com>