Update virtio to latest ABI
This patch updates KVM's virtio implementation to the latest virtio ABI.  This
includes a change in the network header and support for reset.

With this patch, the block and network driver from Rusty's latest queue are
functioning.  Module unload and reload work and I no longer see an error when
repeatedly bringing a network interface up and down.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>