Add libfdt support to qemu
This patch adds libfdt support into the qemu configuation script. There is a
portion for libfdt probing that is has some "XXX". The code commented out is
how to properly do probing for a library, the issue is that with kvm-userspace
we do not build libfdt before configuring qemu. So to pervent this check from
failing everytime we just do a simple test with the libfdt header. Once this
can be resolved elgantly then we can uncomment the place commented out and
remove the lines below them.

Signed-off-by: Jerone Young <jyoung5@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>