Create new load_uimage() & gunzip support to uboot loader
This patch adds the ability for the load address to be caputred when
loading a uImage or cuImage. It also adds a better return code as the
size can be an usigned long. This is done by creating a new function
prototype and calling it load_uimage. To keep compatibility with code
already using the old load_uboot() a wrapper function has been created
so current callers will not break.

Also added is gunzip support to allow for loading of uimages with a
compressed kenrel images.

Signed-off-by: Jerone Young <jyoung5@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>