Ppc: set memory size correctly in the device tree
This fixes an issue where the amount of memory is not properly being
defined in the device tree. It currently is hardcoded for 144MB. The
result is that if you specify a memory size below the hardcoded size,
the guest crashes. This patch now dynamically changes the device tree
to the memory value specified.

Signed-off-by: Jerone Young <jyoung5@us.ibm.com>
Acked-by: Hollis Blanchard <hollisb@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>