Virtio save/restore support
This patch implements the core of save/restore support for virtio.  It's
modelled after how PCI save/restore works.

N.B. This makes savevm/loadvm work, but not live migration.  The issue with
live migration is that we're manipulating guest memory without updating the
dirty bitmap correctly.  I will submit a patch in the near future that
addresses that problem.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>