Remove application name decoration
Decorating the application name is not the best way to let users figure out
if KVM is enabled or not.  It's impossible for management applications to
figure out, inconsistent with other accelerators like kqemu, and clutters the
title bar.

We may experience a little user pain in the short term since we've told users to
look for it to determine if KVM is enabled or not but having a proper monitor
command is better in the long run.

I also think it will be hard to merge this into upstream QEMU whereas a monitor
command will be obvious to merge.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>