Drop the mutex while reading from tapfd
The idea here is that with GSO, packets are much larger
and we can allow the vcpu threads to e.g. process irq
acks during the window where we're reading these
packets from the tapfd.

One known issue with this is that it triggers a subtle
SMP race in the kernel's posix-timers and signalfd code.
See here for more details and a test case:

  http://lkml.org/lkml/2008/7/17/151

The symptoms of this are that:

  a) occassionally throughput drops almost to zero
  b) manually doing "killall -ALRM qemu-kvm" kicks qemu
     out if its funk.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>
Signed-off-by: Avi Kivity <avi@qumranet.com>