kvm: user: define and use AR in configure scripts and Makefiles
libkvm/Makefile uses $(AR), but it's not defined by the top-level configure.
The problem is that when cross-compiling for a 64-bit target on a 32-bit host,
32-bit ar will be used and the 64-bit code will fail to link.

user/Makefile has the same problem, and even uses plain "ar".

Add AR to configure and user/configure, and use it in the Makefiles.

Signed-off-by: Hollis Blanchard <hollisb@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>