Do not access virtio-rings directly
This introduces accessors for all virtio ring members and then uses the
appropriate physical memory accessors.  We shouldn't access guest physical
memory directly and this fixes that.  As a consequence, this helps us get closer
to working live migration.

I've done light performance testing and it seems to not have an impact.  If
further testing demonstrates an impact, we can optimize the physical memory
accessor functions.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>