Do not modify VirtQueueElement
Right now, virtio-net modifies the iovec structure in VirtQueueElement.  This is
evil.  This creates an impossible situation if we want to bounce iovec buffers
in VirtQueueElement as we lose track of the original buffers and the resulting
free results in very bad things.

I tried to refactor receive_headers() and iov_fill() to be able to skip the
header if present but failed miserably.  Instead of spending more time trying to
get that to work, I simply decided to leave the code as-is and copy the iovec
to a temporary buffer.

I doubt there will be any performance impact as usually there are only a few
elements in the iovec anyway.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>