Enable nested SVM support in userspace
In order to use the new nested SVM support in KVM, we need to
expose the SVM capability to the virtual machine.
The current code masks this bit out.

In order to expose the SVM capability to the guest, I added
a -enable-nesting command line option, that makes qemu expose
the SVM capability.
Please bear in mind that this switch does not work properly
on VMX.

So what you need to do to use nested SVM with this patch
applied is:

1. modprobe kvm-amd nested=1
2. qemu-system-x86_64 -enable-nesting ...

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Avi Kivity <avi@redhat.com>