Preallocate -mem-path memory by default
Attached is a patch to qemu which adds an option to preallocate of huge
pages at startup time, making it the default.

The motivation for this arose from odd behavior seen in qemu when access
to huge page backed phys_mem failed during startup (eg: loading the
bios), and during runtime where a guest will terminate via signal if a
free hpage isn't available to satisfy a guest page fault.

This is believed to be an interim solution until proper huge page
support is available from within the kernel+kvm, ideally allowing
fallback to 4K pages on a dynamic basis and ultimately allowing huge
pages faults to reclaim huge pages from other users in the system.

Signed-off-by: john cooper <john.cooper@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>