Fix option_rom_setup_reset address
Commit f2b690ba461971fb8b04354de8717a73fd08b945 changed the target
address for option roms, but failed to use the same address when
registering an option rom reset.  This manifests itself when using
extboot (boot=on) and reseting a guest via reboot or system_reset on
monitor and the guest fails to boot.  This patch register the correct
region for each option rom.

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>