Remove memalign call for guess_disk_lchs
This code doesn't exist in upstream QEMU because it is not necessary to
provide an aligned buffer to bdrv_read.  The API has always worked this way
although at one point, the bouncing was broken which is what led to this
patch.

The places where qemu_memalign() is used in QEMU are only where performance is
sensitive.  guess_disk_lchs does not fall into this category.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>