kvmtrace_format: read /usr/share/kvm/formats by default
This will help distributions to package kvmtrace and kvmtrace_format, making
kvmtrace_format work out-of-the-box.

The directory is defined at the beginning of the script to make it easier for
distributions to customize the path at install-time, depending on packaging
parameters.

There are no install rules to install kvmtrace and kvmtrace_format, but it
would be interesting to add such rules later. They could replace the PREFIX
definition with the prefix set by ./configure.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>