kvm: bios: Move _PR to SSDT
Move _PR block from the DSDT to a new SSDT in the KVM BIOS. As AML,
or at least Windows doesn't allow forward declarations of Methods,
it is necessary to declare the _GPE._L02 Method in the SSDT as well.

This will make it possible to plug in different SSDTs with different
processor counts in the future.

Signed-off-by: Jes Sorensen <jes@sgi.com>
Signed-off-by: Avi Kivity <avi@redhat.com>