libkvm: remove ppc functions from callbacks
handle_powerpc_dcr_read() and handle_powerpc_dcr_write() are two
powerpc specific functions that are called via libkvm callbacks.
However, grepping the source code finds simply no use of them. This
is probably due to the fact that powerpc now relies on a totally
qemu upstream implementation of kvm, and does not need it anymore.

Signed-off-by: Glauber Costa <glommer@redhat.com>
Acked-by: Hollis Blanchard <hollisb@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>