bios: allow qemu to configure irq0->inti2 override
Win2k8 expects the HPET interrupt on inti2, regardless of whether an
override exists in the BIOS. And the HPET spec states that in legacy
mode, timer interrupt is on inti2.

The irq0->inti2 override will always be used unless the kernel cannot
do irq routing (i.e., compatibility with old kernels). So if the
kernel is capable, userspace sets up irq0->inti2 via the irq routing
interface, and adds the irq0->inti2 override to the MADT interrupt
source override table, and the mp table (for the no-acpi case).

Signed-off-by: Beth Kon <eak@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>