HPET support with kvm
The big change here is handling of enabling/disabling of hpet legacy
mode. When hpet enters legacy mode, the spec says that the pit stops
generating interrupts. In practice, we want to stop the pit periodic
timer from running because it is wasteful in a virtual environment.

We also have to worry about the hpet leaving legacy mode (which, at
least in linux, happens only during a shutdown or crash). At this
point, according to the hpet spec, PIT interrupts need to be
reenabled. For us, it means the PIT timer needs to be restarted.

This patch handles this situation better than the earlier versions by
coming closer to just disabling PIT interrupts. It allows the PIT
state to change if the OS modifies it, even while PIT is disabled, but
does not allow a pit timer to start. Then if HPET legacy mode is
disabled, whatever the PIT state is at that point, the PIT timer is
restarted accordingly.

[avi: adjust to kernel header changes]

Signed-off-by: Beth Kon <eak@us.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>