remove env->exit_request usage from qemu-kvm.c
Today I found a very catastrophic regression: I cannot run my mission
critical servers running RHL7.1 anymore. This is a total disaster.

Fortunately, I was able to isolate the commit that caused it:
    commit bb598da496c040d42dde564bd8ace181be52293e
    Author: Glauber Costa <glommer@redhat.com>
    Date:   Mon Jul 6 16:12:52 2009 -0400

This guy is certainly stupid, and deserves punishment. It means I'll
be writting code using emacs for the next week.

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>