configure: make default KVM_VERSION qemu-kvm-devel
Released versions of qemu-kvm now have a KVM_VERSION of either
qemu-kvm-x.y.z or qemu-kvm-devel-XX. However, when building from git,
KVM_VERSION is kvm-devel.

Defaulting to qemu-kvm-devel makes more sense.

With newer versions of qemu-kvm, libvirt actually ignores the kvm
snapshot number (i.e. 86, 87 etc.) and just uses the 'qemu-kvm-'
string to detect that this is the kvm fork of qemu. The actual kvm
snapshot number is no longer important, it's the qemu version number
which is important.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>