Use kvm_run inside CPUState.
There is no need to replicate kvm_run inside vcpu context, since there
is already a field dedicated to that in CPUState. For now, keep both on,
so function can be converted slowly, since a lot of function headers
will be changed to take an env parameter, instead of a vcpu one.

This patch also converts functions that are static to this file as a
first step.

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>